fix: cut homepage to instant by work
