+++ src/pages/about/index.tsx
@@ -6,25 +6,26 @@ export default function AboutIndex() {
       <div className="min-h-screen min-w-screen bg-white flex flex-col items-center text-black">
         <Navbar />
         <main className="flex justify-start lg:justify-center lg:max-w-[980px] mt-[140px] lg:mt-[250px] w-full px-4 mb-40">
-          <div className="flex flex-col gap-16">
+          <div className="flex flex-col gap-5 lg:gap-16">
             {/* div head title */}
             <h1 className="text-[#3C33E6] text-[clamp(32px,_5vw,_60px)] text-inria lg:-mb-10">
               Natthanat <br className="lg:hidden" />
               Chuayriang
             </h1>
-            <h1 className="font-light lg:text-base w-full lg:max-w-[920px]">
+            <h1 className="font-light text-xs lg:text-base w-full lg:max-w-[920px]">
               Lorem Ipsumis simply dummy text of the printing and typesetting
               industry. Lorem Ipsum has been the industrys standard dummy text
               ever since Lorem Ipsumis simply dummy text of the printing and
               typesetting industry. Lorem Ipsum has been the industrys standard
               dummy text{" "}
             </h1>
-            <div className="flex flex-col gap-8 lg:gap-12 lg:mt-28">
+            <div className="flex flex-col gap-8 lg:gap-12 mt-28 lg:mt-20">
               {/* div about me */}
               <div className="flex flex-col gap-3 lg:gap-8">
-                <h1 className="font-semibold lg:text-2xl">ABOUT ME</h1>
+                <h1 className="font-semibold lg:text-2xl">ABOUT ME
+                </h1>
                 <a
-                  href="https://www.canva.com/design/DAGcKCipCd8/jbanjGh2fKuUJzNq7qxJ3g/view?utm_content=DAGcKCipCd8&utm_campaign=designshare&utm_medium=link2&utm_source=uniquelinks&utlId=hc5972c3b4d"
+                  href="/components/resume/Resume.pdf"
                   target="blank"
                   className=" hover:scale-105 hover:text-gray-500 transform origin-left transition-all duration-200 text-sm lg:text-xl flex w-max"
                   style={{ transformOrigin: "left buttom" }}
